Move common dtos to common.go
